apps/api/pyproject.toml
[project]
name = "gp-api"
version = "0.1.0"
description = "Grove Pulse API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
    "sqlalchemy>=2.0,<3",
    "alembic>=1.13",
    "psycopg[binary]>=3.1",
    "argon2-cffi>=23.1",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.26",
    "pytest>=8.0",
    "ruff>=0.4",
]

[project.scripts]
gp-seed = "gp_api.seed.cli:main"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]
